+++ ChangeLog.md
@@ -10,6 +10,7 @@
   ([#84](https://github.com/davep/oldnews/pull/84))
 - Added a `directories` command line command that can be used to see which
   directories are being used by the application.
+  ([#88](https://github.com/davep/oldnews/pull/88))
 
 ## v0.5.0